refactor to support File tranmission (pending)
I want to try sending files this way, and they're too big to go through
in one connection, so there needs to be a converstation about what parts
the receiver has and what parts it needs. First step is to refactor so
the existing code is one subclass that participates in this
scheme. That's done, and works.